moved Geometry:: and ConfigureQC to /configure
