+++ include/dealiiqc/configure/configure_qc.h
@@ -11,8 +11,8 @@
 #include <utility>
 #include <memory>
 
-#include <dealiiqc/io/geometry_box.h>
-#include <dealiiqc/io/geometry_gmsh.h>
+#include <dealiiqc/configure/geometry/geometry_box.h>
+#include <dealiiqc/configure/geometry/geometry_gmsh.h>
 #include <dealiiqc/io/parse_atom_data.h>
 #include <dealiiqc/potentials/pair_lj_cut.h>
 #include <dealiiqc/potentials/pair_coul_wolf.h>
