+++ include/dealiiqc/configure/geometry/geometry_box.h
@@ -5,7 +5,7 @@
 #include <deal.II/grid/grid_generator.h>
 #include <deal.II/grid/grid_tools.h>
 
-#include <dealiiqc/io/geometry_base.h>
+#include <dealiiqc/configure/geometry/geometry_base.h>
 
 namespace dealiiqc
 {
