+++ tests/configure/CMakeLists.txt
@@ -1,0 +1,3 @@
+SET(TEST_LIBRARIES ${TARGETLIB})
+SET(TEST_TARGET ${TARGET})
+DEAL_II_PICKUP_TESTS()
