+++ tests/configure/geometry_box_01.output
